cleaned up code and created Form class for some methods and to use as a model
