--- PDFFiller/Models/Form.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PDFFiller.Models;
-
-internal class Form
-{
-    public string? HereAfter { get; set; }
-    public string? TwoZeroTwo { get; set; }
-    public string? Date { get; set; }
-    public string? Name { get; set; }
-    public string? Title { get; set; }
-    public string? CompanyName { get; set; }
-    public string? Date_2 { get; set; }
-}
